separate bucket creation and main
