+++ main.py
@@ -120,70 +120,6 @@ class MyTableModel(QtCore.QAbstractTableModel):
         return True
 
 
-
-
-
-
-
-
-######################################################################################################################################
-####################### BUCKET CREATE UI ##################################
-
-class BucketCreateUI(QtGui.QMainWindow):
-    def __init__(self, parent=None):
-        QtGui.QWidget.__init__(self, parent)
-        self.bucket_create_ui = Ui_BucketCreate()
-        self.bucket_create_ui.setupUi(self)
-
-        QtCore.QObject.connect(self.bucket_create_ui.create_bucket_bt, QtCore.SIGNAL("clicked()"),
-                               self.createNewBucketCreateThread)  # create bucket action
-        QtCore.QObject.connect(self.bucket_create_ui.cancel_bt, QtCore.SIGNAL("clicked()"),
-                               self.close)  # create bucket action
-
-        self.connect(self, SIGNAL("showBucketCreatingException"), self.show_bucket_creating_exception_dialog)
-        self.connect(self, SIGNAL("showBucketCreatedSuccessfully"), self.show_bucket_crated_successfully)
-        self.connect(self, SIGNAL("showBucketCreationMissingFields"), lambda : QMessageBox.about(self, "Warning", "Please fill out all fields!"))
-
-        self.storj_engine = StorjEngine()  # init StorjEngine
-
-    def show_bucket_creating_exception_dialog(self, exception):
-        QMessageBox.about(self, "Unhandled exception while creating bucket", "Exception: " + str(exception))
-
-    def show_bucket_crated_successfully(self, bucket_name):
-        msgBox = QtGui.QMessageBox(QtGui.QMessageBox.Information, "Success", "Bucket '" + str(bucket_name) + "' was created successfully!",
-                                   QtGui.QMessageBox.Ok)
-        msgBox.exec_()
-
-    def createNewBucketCreateThread(self):
-        bucket_create_thread = threading.Thread(target=self.create_bucket, args=())
-        bucket_create_thread.start()
-
-    def create_bucket(self):
-        self.bucket_name = self.bucket_create_ui.bucket_name.text()
-        self.bucket_storage = self.bucket_create_ui.bucket_storage_size.text()
-        self.bucket_transfer = self.bucket_create_ui.bucket_transfer.text()
-
-        bucekt_cerated = False  # init boolean
-        if self.bucket_name != "" and self.bucket_transfer != "" and self.bucket_storage != "":
-
-            try:
-                self.storj_engine.storj_client.bucket_create(str(self.bucket_name), int(self.bucket_storage),
-                                                             int(self.bucket_transfer))
-                bucekt_cerated = True
-            except  storj.exception.StorjBridgeApiError, e:
-                bucekt_cerated = False
-                self.emit(SIGNAL("showBucketCreatingException"), str(e))
-
-        else:
-            self.emit(SIGNAL("showBucketCreationMissingFields"))
-            bucekt_cerated = False
-
-        if bucekt_cerated:
-            self.emit(SIGNAL("showBucketCreatedSuccessfully"), str(self.bucket_name))  # show dialog - bucket created successfully
-
-        print 1
-
-
 ######################################################################################################################################
 ####################### FILE MANAGER UI ##################################
